Fixes and improvements
Improved session handling, with session regeneration when a sucessful
login occour, and after 10 times the SESSION_TIMEOUT, make session more
robust against session hijack/fixation.
Sesion expiration are now better handled, timeout happen between the
finish of a page processing and the next access;

Session is now enforced rightly when an AJAX operation is called.

Issue #45, unescaped output on TopRules chart, make chart unusable
with special character in rule description. Thanks to Ibere Tizio,
by report and propose a fix.

Change the management menu behavior, to address a performance impact
described by Juraj Sakala (in issue #14 on GitHub). A temporary
denial of service happens while database events are counted for get
sensor information. Now the user can change (on config.php and in
runtime) if Sensor Info will be loaded or not.
Database schema change to fix this, but will change in update process
only on next release

In some circunstances the filter dialog are unable to show web
hostnames (ie. when more that 2000 entries are available), this
is a limitation/protection of browsers. Now the hostname was
changed from a combo-box, to a text input with autocomplete powered
by a Ajax interface, that brings hostnames when you type it on
hostname field. This fixes issue #15 on github. Thanks to Juraj
Sakala for report and discuss it;